+++ src/AABB.h
@@ -13,7 +13,7 @@ public:
 	void ComputeFromVertices(const std::vector<Vertex>& vertices);
 	void Recompute(const XMFLOAT4X4& matrix);
 
-	MSWRL::ComPtr<ID3D12Resource> CreateBuffer(UINT64 size, D3D12_HEAP_TYPE heapType, D3D12_RESOURCE_STATES initialState);
+	MSWRL::ComPtr<ID3D12Resource> CreateBuffer(uint64_t size, D3D12_HEAP_TYPE heapType, D3D12_RESOURCE_STATES initialState);
 	void UploadBuffers();
 
 	void BindMeshData(MSWRL::ComPtr<ID3D12GraphicsCommandList> commandList);
@@ -32,5 +32,5 @@ private:
 	std::vector<uint32_t> _aabbIndices;
 	MSWRL::ComPtr<ID3D12Resource> _indexBuffer;
 	D3D12_INDEX_BUFFER_VIEW _indexBufferView = {};
-	size_t _indicesSize;
+	int32_t _indicesSize;
 };
